requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
argcomplete==2.0.0
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
boto3==1.26.68
botocore==1.29.68
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-s3-storage==0.13.11
djangorestframework==3.14.0
durationpy==0.5
environ==1.0
frozenlist==1.3.3
gunicorn==20.1.0
hjson==3.1.0
idna==3.4
jmespath==1.0.1
jsonschema==4.17.3
kappa==0.6.0
MarkupSafe==2.1.2
multidict==6.0.4
mysqlclient==2.1.1
placebo==0.9.0
PyMySQL==1.0.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.3.0
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
zappa==0.56.1